+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 + **[BC BREAK] Changed signature and response of ApiAdminStorage::fileUpload() and ApiAdminStorage::imageUpload().** The methods return now the desired object instead of a PostResponse object.
 + Removed deprecated methods in ApiAdminStorage
 + Removed deprecated classes in `luya\headless\modules\models\*` folder.
++ [#17](https://github.com/luyadev/luya-headless/issues/17) Fixed bug where endpointName token was unable to override.
 
 ## 1.2.0 (19. March 2019)